PythonServices/requirements.txt
flask
flask-cors
face_recognition
pillow
numpy
